App fcm updated (#39)
* agent_close_read_received_events_integrated

* reconnection_limit_and_branding_mapping_changes

* Update README.md

* reconnection_limit_and_branding_new_changes

* exposed_config_dynamic_timestamp_bot_icon_display

* token_changes

* app fcm updated
