+++ BotsSDK/app/.gitignore
@@ -1,1 +1,2 @@
 /build
+/google-services.json
