+++ BotsSDK/korebot/.gitignore
@@ -1,2 +1,1 @@
 /build
-/google-services.json
