+++ BotsSDK/korebot/build.gradle
@@ -1,10 +1,9 @@
 apply plugin: 'com.android.library'
-//apply plugin: 'com.google.gms.google-services'
 
 android {
     compileSdk 34
 
-    dataBinding{
+    dataBinding {
         enabled = true
     }
     defaultConfig {
@@ -43,7 +42,7 @@ dependencies {
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
     implementation 'de.greenrobot:eventbus:2.4.0'
 
-    implementation(project(':korebotsdklib')){
+    implementation(project(':korebotsdklib')) {
         exclude module: 'support-v4'
     }
 
@@ -54,7 +53,7 @@ dependencies {
     implementation 'com.google.code.gson:gson:2.10.1'
     implementation 'androidx.cardview:cardview:1.0.0'
     implementation 'androidx.recyclerview:recyclerview:1.3.2'
-    implementation ('net.gotev:speech:1.6.2'){
+    implementation('net.gotev:speech:1.6.2') {
         exclude module: 'support-v4'
     }
 
@@ -69,9 +68,6 @@ dependencies {
     implementation 'com.github.bumptech.glide:glide:4.16.0'
     implementation 'com.github.emojione:emojione-android:3.1.0'
     implementation 'com.github.cachapa:ExpandableLayout:2.9.2'
-    //Firebase Cloud Messaging
-    implementation 'com.google.firebase:firebase-core:21.1.1'
-    implementation 'com.google.firebase:firebase-messaging:24.0.0'
 
 //    implementation 'org.apache.httpcomponents:httpclient:4.5.3'
 //    implementation 'org.apache.httpcomponents:httpcore:4.4.6'
