+++ BotsSDK/korebot/src/main/java/kore/botssdk/application/BotApplication.java
@@ -5,14 +5,10 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.net.ConnectivityManager;
 
-import com.google.firebase.FirebaseApp;
-
-import kore.botssdk.fcm.FCMWrapper;
 import kore.botssdk.listener.NetworkStateReceiver;
 import kore.botssdk.utils.ClosingService;
 
 /**
- * Created by Pradeep Mahato on 31-May-16.
  * Copyright (c) 2014 Kore Inc. All rights reserved.
  */
 public class BotApplication extends Application {
@@ -23,8 +19,6 @@ public class BotApplication extends Application {
     public void onCreate() {
         super.onCreate();
         appControl = new AppControl(getApplicationContext());
-        FirebaseApp.initializeApp(getApplicationContext());
-        FCMWrapper.getInstance().init();
         IntentFilter filter = new IntentFilter(ConnectivityManager.CONNECTIVITY_ACTION);
         registerReceiver(new NetworkStateReceiver(), filter);
         startService(new Intent(getApplicationContext(), ClosingService.class));
